add MVars
MVars are mostly implemented using the definition from Claessen 1999, using their concurrency model. IO and IORefs (non-concurrent mutable variables) are postulated in Agda, while the Haskell code uses the builtin ones.
Definition of writeMVar is adjusted to correspond to actual Haskell implementation, according to which it blocks when the MVar is full.